use uint128 to save on storage slots

+++ src/libraries/LinkedListQueue.sol
@@ -5,35 +5,35 @@ import {console} from "forge-std/Test.sol";
 
 library LinkedListQueue {
     struct Node {
-        uint256 prev;
-        uint256 next;
+        uint128 prev;
+        uint128 next;
     }
 
     struct Queue {
         mapping(uint256 => Node) nodes;
-        uint256 head;
-        uint256 tail;
-        uint256 length;
+        uint128 head;
+        uint128 tail;
+        uint128 length;
     }
 
     /// @notice Error for token that does not exist.
-    error TokenDoesNotExist(uint256 id);
+    error TokenDoesNotExist(uint128 id);
 
     /// @notice Checks if _id is in the queue.
-    function exists(Queue storage q, uint256 _id) internal view returns (bool) {
+    function exists(Queue storage q, uint128 _id) internal view returns (bool) {
         return q.nodes[_id].prev != 0 || q.nodes[_id].next != 0 || _id == q.head || _id == q.tail;
     }
 
     /// @notice Returns the node with _id in O(1).
     /// @dev Reverts if _id is not in the queue.
-    function at(Queue storage q, uint256 _id) internal view returns (Node storage n) {
+    function at(Queue storage q, uint128 _id) internal view returns (Node storage n) {
         require(exists(q, _id), TokenDoesNotExist(_id));
         n = q.nodes[_id];
     }
 
     /// @notice Pushes a new node to the front.
     /// @dev Reverts if _id is zero or already in use.
-    function pushFront(Queue storage q, uint256 _id) internal {
+    function pushFront(Queue storage q, uint128 _id) internal {
         require(_id != 0, "Invalid id");
         require(!exists(q, _id), "ID already used");
 
@@ -53,7 +53,7 @@ library LinkedListQueue {
 
     /// @notice Pushes a new node to the back.
     /// @dev Reverts if _id is zero or already in use.
-    function pushBack(Queue storage q, uint256 _id) internal {
+    function pushBack(Queue storage q, uint128 _id) internal {
         require(_id != 0, "Invalid id");
         require(!exists(q, _id), "ID already used");
 
@@ -73,7 +73,7 @@ library LinkedListQueue {
 
     /// @notice Inserts a new node immediately before an existing node _cursor.
     /// @dev Reverts if _id is zero or already used, or if _cursor doesn't exist.
-    function insertBefore(Queue storage q, uint256 _cursor, uint256 _id) internal {
+    function insertBefore(Queue storage q, uint128 _cursor, uint128 _id) internal {
         require(exists(q, _cursor), "Cursor does not exist");
         require(_id != 0, "Invalid id");
         require(!exists(q, _id), "ID already used");
@@ -97,7 +97,7 @@ library LinkedListQueue {
 
     /// @notice Inserts a new node immediately after an existing node _cursor.
     /// @dev Reverts if _id is zero or already used, or if _cursor doesn't exist.
-    function insertAfter(Queue storage q, uint256 _cursor, uint256 _id) internal {
+    function insertAfter(Queue storage q, uint128 _cursor, uint128 _id) internal {
         require(exists(q, _cursor), "Cursor does not exist");
         require(_id != 0, "Invalid id");
         require(!exists(q, _id), "ID already used");
@@ -121,7 +121,7 @@ library LinkedListQueue {
 
     /// @notice Removes the node with id `_id` in O(1).
     /// @dev Reverts if queue is empty or `_id` is invalid.
-    function remove(Queue storage q, uint256 _id) internal {
+    function remove(Queue storage q, uint128 _id) internal {
         require(q.length > 0, "Empty");
 
         Node storage n = at(q, _id);
